+++ src/Products/PageTemplates/Extensions/find_bad_templates.py
@@ -33,7 +33,9 @@ ERROR_HTML = """
 def find_bad_templates(self):
     pt_errors = {}
     html_output = ''
-    pts = self.ZopeFind(self, obj_metatypes=('Page Template',),
+    pts = self.ZopeFind(self,
+                        obj_metatypes=('Page Template',
+                                       'Filesystem Page Template'),
                         search_sub=True)
 
     for (pt_path, pt) in pts:
@@ -43,7 +45,9 @@ def find_bad_templates(self):
         try:
             pt.pt_macros()
         except PTRuntimeError:
-            pt_errors[pt_path] = pt._v_errors
+            # html quote "<" characters to be displayed as such
+            errs = [err.replace('<', '&lt;') for err in pt._v_errors]
+            pt_errors[pt_path] = errs
 
     for pt_path in sorted(pt_errors.keys()):
         html_output += ERROR_HTML % {'pt_path': pt_path,